+++ service/kafkaconnect/api.go
@@ -436,6 +436,109 @@ func (c *KafkaConnect) DeleteConnectorWithContext(ctx aws.Context, input *Delete
 	return out, req.Send()
 }
 
+const opDeleteCustomPlugin = "DeleteCustomPlugin"
+
+// DeleteCustomPluginRequest generates a "aws/request.Request" representing the
+// client's request for the DeleteCustomPlugin operation. The "output" return
+// value will be populated with the request's response once the request completes
+// successfully.
+//
+// Use "Send" method on the returned Request to send the API call to the service.
+// the "output" return value is not valid until after Send returns without error.
+//
+// See DeleteCustomPlugin for more information on using the DeleteCustomPlugin
+// API call, and error handling.
+//
+// This method is useful when you want to inject custom logic or configuration
+// into the SDK's request lifecycle. Such as custom headers, or retry logic.
+//
+//
+//    // Example sending a request using the DeleteCustomPluginRequest method.
+//    req, resp := client.DeleteCustomPluginRequest(params)
+//
+//    err := req.Send()
+//    if err == nil { // resp is now filled
+//        fmt.Println(resp)
+//    }
+//
+// See also, https://docs.aws.amazon.com/goto/WebAPI/kafkaconnect-2021-09-14/DeleteCustomPlugin
+func (c *KafkaConnect) DeleteCustomPluginRequest(input *DeleteCustomPluginInput) (req *request.Request, output *DeleteCustomPluginOutput) {
+	op := &request.Operation{
+		Name:       opDeleteCustomPlugin,
+		HTTPMethod: "DELETE",
+		HTTPPath:   "/v1/custom-plugins/{customPluginArn}",
+	}
+
+	if input == nil {
+		input = &DeleteCustomPluginInput{}
+	}
+
+	output = &DeleteCustomPluginOutput{}
+	req = c.newRequest(op, input, output)
+	return
+}
+
+// DeleteCustomPlugin API operation for Managed Streaming for Kafka Connect.
+//
+// Deletes a custom plugin.
+//
+// Returns awserr.Error for service API and SDK errors. Use runtime type assertions
+// with awserr.Error's Code and Message methods to get detailed information about
+// the error.
+//
+// See the AWS API reference guide for Managed Streaming for Kafka Connect's
+// API operation DeleteCustomPlugin for usage and error information.
+//
+// Returned Error Types:
+//   * NotFoundException
+//   HTTP Status Code 404: Resource not found due to incorrect input. Correct
+//   your request and then retry it.
+//
+//   * BadRequestException
+//   HTTP Status Code 400: Bad request due to incorrect input. Correct your request
+//   and then retry it.
+//
+//   * ForbiddenException
+//   HTTP Status Code 403: Access forbidden. Correct your credentials and then
+//   retry your request.
+//
+//   * ServiceUnavailableException
+//   HTTP Status Code 503: Service Unavailable. Retrying your request in some
+//   time might resolve the issue.
+//
+//   * TooManyRequestsException
+//   HTTP Status Code 429: Limit exceeded. Resource limit reached.
+//
+//   * UnauthorizedException
+//   HTTP Status Code 401: Unauthorized request. The provided credentials couldn't
+//   be validated.
+//
+//   * InternalServerErrorException
+//   HTTP Status Code 500: Unexpected internal server error. Retrying your request
+//   might resolve the issue.
+//
+// See also, https://docs.aws.amazon.com/goto/WebAPI/kafkaconnect-2021-09-14/DeleteCustomPlugin
+func (c *KafkaConnect) DeleteCustomPlugin(input *DeleteCustomPluginInput) (*DeleteCustomPluginOutput, error) {
+	req, out := c.DeleteCustomPluginRequest(input)
+	return out, req.Send()
+}
+
+// DeleteCustomPluginWithContext is the same as DeleteCustomPlugin with the addition of
+// the ability to pass a context and additional request options.
+//
+// See DeleteCustomPlugin for details on how to use this API operation.
+//
+// The context must be non-nil and will be used for request cancellation. If
+// the context is nil a panic will occur. In the future the SDK may create
+// sub-contexts for http.Requests. See https://golang.org/pkg/context/
+// for more information on using Contexts.
+func (c *KafkaConnect) DeleteCustomPluginWithContext(ctx aws.Context, input *DeleteCustomPluginInput, opts ...request.Option) (*DeleteCustomPluginOutput, error) {
+	req, out := c.DeleteCustomPluginRequest(input)
+	req.SetContext(ctx)
+	req.ApplyOptions(opts...)
+	return out, req.Send()
+}
+
 const opDescribeConnector = "DescribeConnector"
 
 // DescribeConnectorRequest generates a "aws/request.Request" representing the
@@ -2311,8 +2414,12 @@ type CreateConnectorInput struct {
 
 	// A map of keys to values that represent the configuration for the connector.
 	//
+	// ConnectorConfiguration is a sensitive parameter and its value will be
+	// replaced with "sensitive" in string returned by CreateConnectorInput's
+	// String and GoString methods.
+	//
 	// ConnectorConfiguration is a required field
-	ConnectorConfiguration map[string]*string `locationName:"connectorConfiguration" type:"map" required:"true"`
+	ConnectorConfiguration map[string]*string `locationName:"connectorConfiguration" type:"map" required:"true" sensitive:"true"`
 
 	// A summary description of the connector.
 	ConnectorDescription *string `locationName:"connectorDescription" type:"string"`
@@ -2745,8 +2852,12 @@ type CreateWorkerConfigurationInput struct {
 
 	// Base64 encoded contents of connect-distributed.properties file.
 	//
+	// PropertiesFileContent is a sensitive parameter and its value will be
+	// replaced with "sensitive" in string returned by CreateWorkerConfigurationInput's
+	// String and GoString methods.
+	//
 	// PropertiesFileContent is a required field
-	PropertiesFileContent *string `locationName:"propertiesFileContent" type:"string" required:"true"`
+	PropertiesFileContent *string `locationName:"propertiesFileContent" type:"string" required:"true" sensitive:"true"`
 }
 
 // String returns the string representation.
@@ -3349,6 +3460,96 @@ func (s *DeleteConnectorOutput) SetConnectorState(v string) *DeleteConnectorOutp
 	return s
 }
 
+type DeleteCustomPluginInput struct {
+	_ struct{} `type:"structure" nopayload:"true"`
+
+	// The Amazon Resource Name (ARN) of the custom plugin that you want to delete.
+	//
+	// CustomPluginArn is a required field
+	CustomPluginArn *string `location:"uri" locationName:"customPluginArn" type:"string" required:"true"`
+}
+
+// String returns the string representation.
+//
+// API parameter values that are decorated as "sensitive" in the API will not
+// be included in the string output. The member name will be present, but the
+// value will be replaced with "sensitive".
+func (s DeleteCustomPluginInput) String() string {
+	return awsutil.Prettify(s)
+}
+
+// GoString returns the string representation.
+//
+// API parameter values that are decorated as "sensitive" in the API will not
+// be included in the string output. The member name will be present, but the
+// value will be replaced with "sensitive".
+func (s DeleteCustomPluginInput) GoString() string {
+	return s.String()
+}
+
+// Validate inspects the fields of the type to determine if they are valid.
+func (s *DeleteCustomPluginInput) Validate() error {
+	invalidParams := request.ErrInvalidParams{Context: "DeleteCustomPluginInput"}
+	if s.CustomPluginArn == nil {
+		invalidParams.Add(request.NewErrParamRequired("CustomPluginArn"))
+	}
+	if s.CustomPluginArn != nil && len(*s.CustomPluginArn) < 1 {
+		invalidParams.Add(request.NewErrParamMinLen("CustomPluginArn", 1))
+	}
+
+	if invalidParams.Len() > 0 {
+		return invalidParams
+	}
+	return nil
+}
+
+// SetCustomPluginArn sets the CustomPluginArn field's value.
+func (s *DeleteCustomPluginInput) SetCustomPluginArn(v string) *DeleteCustomPluginInput {
+	s.CustomPluginArn = &v
+	return s
+}
+
+type DeleteCustomPluginOutput struct {
+	_ struct{} `type:"structure"`
+
+	// The Amazon Resource Name (ARN) of the custom plugin that you requested to
+	// delete.
+	CustomPluginArn *string `locationName:"customPluginArn" type:"string"`
+
+	// The state of the custom plugin.
+	CustomPluginState *string `locationName:"customPluginState" type:"string" enum:"CustomPluginState"`
+}
+
+// String returns the string representation.
+//
+// API parameter values that are decorated as "sensitive" in the API will not
+// be included in the string output. The member name will be present, but the
+// value will be replaced with "sensitive".
+func (s DeleteCustomPluginOutput) String() string {
+	return awsutil.Prettify(s)
+}
+
+// GoString returns the string representation.
+//
+// API parameter values that are decorated as "sensitive" in the API will not
+// be included in the string output. The member name will be present, but the
+// value will be replaced with "sensitive".
+func (s DeleteCustomPluginOutput) GoString() string {
+	return s.String()
+}
+
+// SetCustomPluginArn sets the CustomPluginArn field's value.
+func (s *DeleteCustomPluginOutput) SetCustomPluginArn(v string) *DeleteCustomPluginOutput {
+	s.CustomPluginArn = &v
+	return s
+}
+
+// SetCustomPluginState sets the CustomPluginState field's value.
+func (s *DeleteCustomPluginOutput) SetCustomPluginState(v string) *DeleteCustomPluginOutput {
+	s.CustomPluginState = &v
+	return s
+}
+
 type DescribeConnectorInput struct {
 	_ struct{} `type:"structure" nopayload:"true"`
 
@@ -3409,7 +3610,11 @@ type DescribeConnectorOutput struct {
 	ConnectorArn *string `locationName:"connectorArn" type:"string"`
 
 	// A map of keys to values that represent the configuration for the connector.
-	ConnectorConfiguration map[string]*string `locationName:"connectorConfiguration" type:"map"`
+	//
+	// ConnectorConfiguration is a sensitive parameter and its value will be
+	// replaced with "sensitive" in string returned by DescribeConnectorOutput's
+	// String and GoString methods.
+	ConnectorConfiguration map[string]*string `locationName:"connectorConfiguration" type:"map" sensitive:"true"`
 
 	// A summary description of the connector.
 	ConnectorDescription *string `locationName:"connectorDescription" type:"string"`
@@ -3449,6 +3654,9 @@ type DescribeConnectorOutput struct {
 	// The Amazon Resource Name (ARN) of the IAM role used by the connector to access
 	// Amazon Web Services resources.
 	ServiceExecutionRoleArn *string `locationName:"serviceExecutionRoleArn" type:"string"`
+
+	// Details about the state of a connector.
+	StateDescription *StateDescription `locationName:"stateDescription" type:"structure"`
 
 	// Specifies which worker configuration was used for the connector.
 	WorkerConfiguration *WorkerConfigurationDescription `locationName:"workerConfiguration" type:"structure"`
@@ -3562,6 +3770,12 @@ func (s *DescribeConnectorOutput) SetServiceExecutionRoleArn(v string) *Describe
 	return s
 }
 
+// SetStateDescription sets the StateDescription field's value.
+func (s *DescribeConnectorOutput) SetStateDescription(v *StateDescription) *DescribeConnectorOutput {
+	s.StateDescription = v
+	return s
+}
+
 // SetWorkerConfiguration sets the WorkerConfiguration field's value.
 func (s *DescribeConnectorOutput) SetWorkerConfiguration(v *WorkerConfigurationDescription) *DescribeConnectorOutput {
 	s.WorkerConfiguration = v
@@ -3638,6 +3852,9 @@ type DescribeCustomPluginOutput struct {
 
 	// The name of the custom plugin.
 	Name *string `locationName:"name" type:"string"`
+
+	// Details about the state of a custom plugin.
+	StateDescription *StateDescription `locationName:"stateDescription" type:"structure"`
 }
 
 // String returns the string representation.
@@ -3691,6 +3908,12 @@ func (s *DescribeCustomPluginOutput) SetLatestRevision(v *CustomPluginRevisionSu
 // SetName sets the Name field's value.
 func (s *DescribeCustomPluginOutput) SetName(v string) *DescribeCustomPluginOutput {
 	s.Name = &v
+	return s
+}
+
+// SetStateDescription sets the StateDescription field's value.
+func (s *DescribeCustomPluginOutput) SetStateDescription(v *StateDescription) *DescribeCustomPluginOutput {
+	s.StateDescription = v
 	return s
 }
 
@@ -5571,6 +5794,47 @@ func (s *ServiceUnavailableException) RequestID() string {
 	return s.RespMetadata.RequestID
 }
 
+// Details about the state of a resource.
+type StateDescription struct {
+	_ struct{} `type:"structure"`
+
+	// A code that describes the state of a resource.
+	Code *string `locationName:"code" type:"string"`
+
+	// A message that describes the state of a resource.
+	Message *string `locationName:"message" type:"string"`
+}
+
+// String returns the string representation.
+//
+// API parameter values that are decorated as "sensitive" in the API will not
+// be included in the string output. The member name will be present, but the
+// value will be replaced with "sensitive".
+func (s StateDescription) String() string {
+	return awsutil.Prettify(s)
+}
+
+// GoString returns the string representation.
+//
+// API parameter values that are decorated as "sensitive" in the API will not
+// be included in the string output. The member name will be present, but the
+// value will be replaced with "sensitive".
+func (s StateDescription) GoString() string {
+	return s.String()
+}
+
+// SetCode sets the Code field's value.
+func (s *StateDescription) SetCode(v string) *StateDescription {
+	s.Code = &v
+	return s
+}
+
+// SetMessage sets the Message field's value.
+func (s *StateDescription) SetMessage(v string) *StateDescription {
+	s.Message = &v
+	return s
+}
+
 // HTTP Status Code 429: Limit exceeded. Resource limit reached.
 type TooManyRequestsException struct {
 	_            struct{}                  `type:"structure"`
@@ -6036,7 +6300,11 @@ type WorkerConfigurationRevisionDescription struct {
 	Description *string `locationName:"description" type:"string"`
 
 	// Base64 encoded contents of the connect-distributed.properties file.
-	PropertiesFileContent *string `locationName:"propertiesFileContent" type:"string"`
+	//
+	// PropertiesFileContent is a sensitive parameter and its value will be
+	// replaced with "sensitive" in string returned by WorkerConfigurationRevisionDescription's
+	// String and GoString methods.
+	PropertiesFileContent *string `locationName:"propertiesFileContent" type:"string" sensitive:"true"`
 
 	// The description of a revision of the worker configuration.
 	Revision *int64 `locationName:"revision" type:"long"`